+++ Changes.md
@@ -20,6 +20,7 @@ API
 ---
 
 - LayoutPlugValueWidget : Added support for `layoutPlugValueWidget:orientation` metadata, with values of either `vertical` or `horizontal`.
+- SpreadsheetUI : Added `addRowButtonMenuSignal` to allow customisation of Spreadsheet row addition within the UI (#4123).
 
 0.59.2.0 (relative to 0.59.1.0)
 ========